requirements.txt
faiss-cpu
numpy
sentence-transformers
transformers
huggingface-hub